+++ src/util/unix.hpp
@@ -49,7 +49,7 @@ TError SetOomScoreAdj(int value);
 TError Run(const std::vector<std::string> &command, int &status, bool stdio = false);
 TError AllocLoop(const TPath &path, size_t size);
 TError Popen(const std::string &cmd, std::vector<std::string> &lines);
-size_t GetNumCores();
+int GetNumCores();
 TError PackTarball(const TPath &tar, const TPath &path);
 TError UnpackTarball(const TPath &tar, const TPath &path);
 TError CopyRecursive(const TPath &src, const TPath &dst);
